Feat: 게시글 권한 기능 추가 - 로그인 사용자 게시글 작성 권한 부여 - 작성자 닉네임, 수정/삭제 선택 부여 기능 추가

+++ diary/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
 
 from.models import Diary
 from .forms import DiaryForm
@@ -20,9 +21,11 @@ def Detail(request, diary_id):
 ## CRUD
 
 #Create
+@login_required()
 def New(request): 
     return render(request, 'diary/edit.html', {'submitType': "발행"})
 
+@login_required()
 def Create(request): 
     #POST Request
     if request.method == "POST":
@@ -30,6 +33,8 @@ def Create(request):
         
         if form.is_valid():
             diary = form.save(commit=False)
+            diary.writer = request.user
+            
             diary.save()
             return redirect('diary:index')
         
@@ -49,27 +54,36 @@ def Edit(request):
 def Update(request, diary_id):
     diary = get_object_or_404(Diary, pk=diary_id) #content upload
 
-    #POST Request
-    if request.method == "POST":
-        form = DiaryForm(request.POST, instance=diary)
+    if request.user == diary.writer:
+        #POST Request
+        if request.method == "POST":
+            form = DiaryForm(request.POST, instance=diary)
+            
+            if form.is_valid():
+                diary = form.save(commit=False)
+                diary.save()
+                return redirect('diary:detail', diary_id=diary.pk)
+            
+            else:
+                return redirect('diary:index')
         
-        if form.is_valid():
-            diary = form.save(commit=False)
-            diary.save()
-            return redirect('diary:detail', diary_id=diary.pk)
-        
+        #GET Request
         else:
-           return redirect('diary:index')
-    
-    #GET Request
+            form = DiaryForm(instance=diary)
+            return render(request, 'diary/edit.html', {'form' : form, 'submitType': "수정"}) 
+        
     else:
-        form = DiaryForm(instance=diary)
-        return render(request, 'edit.html', {'form' : form, 'submitType': "수정"}) 
+        return redirect('diary:index')
 
 
 #Delete
 def Delete(request, diary_id):
     diary = get_object_or_404(Diary, pk=diary_id)
-    diary.delete()
 
-    return redirect('diary:index')
+    # 사용자 본인 확인
+    if request.user == diary.writer: 
+        diary.delete()
+        return redirect('diary:index')
+    
+    else:
+        return redirect('diary:index')
